% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{affinityMatrix}
\alias{affinityMatrix}
\title{This is the affinity Matrix function extracted from SNFtool package.
Computes affinity matrix from a generic distance matrix}
\usage{
affinityMatrix(Diff, K = 20, sigma = 0.5)
}
\arguments{
\item{Diff}{Distance matrix}

\item{K}{Number of nearest neighbors}

\item{sigma}{Variance for local model}
}
\value{
Returns an affinity matrix that represents the neighborhood graph of the data points.
}
\description{
This is the affinity Matrix function extracted from SNFtool package.
Computes affinity matrix from a generic distance matrix
}
\examples{
data(miRNAExp)
Dist1=dist2(t(miRNAExp),t(miRNAExp))
W1 = affinityMatrix(Dist1, 20, 0.5)

}
