% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{data.checkDistribution}
\alias{data.checkDistribution}
\title{Data check distribution}
\usage{
data.checkDistribution(Data)
}
\arguments{
\item{Data}{A matrix representing the genomic data such as gene expression data, miRNA expression data.\cr
For the matrix, the rows represent the genomic features, and the columns represent the samples.}
}
\value{
A plot describes the mean, variance and Median Absolute Deviation (MAD) distribution of features.
}
\description{
Data check distribution
}
\examples{
data(GeneExp)
data.checkDistribution(GeneExp)
}

