% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{data.normalization}
\alias{data.normalization}
\title{Data normalization}
\usage{
data.normalization(Data, type = "feature_Median", log2 = FALSE)
}
\arguments{
\item{Data}{A matrix representing the genomic data such as gene expression data, miRNA expression data. \cr
For the matrix, the rows represent the genomic features, and the columns represent the samples.}

\item{type}{A character value representing the normalization type. The optional values are shown below:
\itemize{
\item "feature_Median". The default value. Normalize dataset by sweeping the median values of each feature.
\item "feature_Mean". Normalize dataset by sweeping the mean values of each feature.    
\item "feature_zscore". Conduct z_score normalization for each feature.
\item "sample_zscore".  Conduct z_score normalization for each samples.

}}

\item{log2}{A logical value. If TRUE, the data is transform as log2(x+1). This is commonly used for RNAseq data.}
}
\value{
The normalized data matrix.
}
\description{
Conduct normalization for dataset.
}
\examples{
data(GeneExp)
result=data.normalization(GeneExp,type="feature_Median",log2=FALSE)

}

