% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valiation.R
\name{drawHeatmap}
\alias{drawHeatmap}
\title{Generate heatmaps}
\usage{
drawHeatmap(data, group = NULL, silhouette = NULL, scale = "no",
  labRow = NULL, labCol = NULL, color = colorRampPalette(c("green",
  "black", "red"))(300), Title = NA)
}
\arguments{
\item{data}{A matrix representing the genomic data such as gene expression data, miRNA expression data.\cr
For the matrix, the rows represent the genomic features, and the columns represent the samples.}

\item{group}{A vector representing the subtype on each sample.  The default is NULL. If it is not
NULL, the samples will be rearrangement according to the subtypes in the heatmap.}

\item{silhouette}{An object of class silhouette. It is a result from function 
silhouette() or silhouette_SimilarityMatrix(). The default is NULL. If it is not NULL,  an annotation will be drawn to show the silhouette width for each sample.}

\item{scale}{A string for data normalization type before heatmap drawing.
The optional values are shown below:
\itemize{
\item "no". No normalization. This is default.
\item "z_score". Normalize data by z_score of features. 
\item "max_min". Normalize each feature by (value-min)/(max-min).
}}

\item{labRow}{labels for the rows. Possible values are:
\itemize{
\item NULL. The default value. It will use the row names of the matrix for the heatmap labels.
\item NA. No row label will be shown. 
\item A list of labels.
}}

\item{labCol}{labels for the columns.  See labRow.}

\item{color}{color specification for the heatmap.}

\item{Title}{A string for the Main title of the heatmap.}
}
\value{
A heatmap
}
\description{
Generate heatmap for datasets.
}
\details{
We applied the R package "NMF" function "aheatmap()" as the heatmap drawer.
}
\examples{
### SNF result analysis
data(GeneExp)
data(miRNAExp)
data(time)
data(status)
GBM=list(GeneExp=GeneExp,miRNAExp=miRNAExp)
result=ExecuteSNF(GBM, clusterNum=3, K=20, alpha=0.5, t=20)
group=result$group
distanceMatrix=result$distanceMatrix
silhouette=silhouette_SimilarityMatrix(group, distanceMatrix)
drawHeatmap(GeneExp,group,silhouette=silhouette,scale="max_min",Title="GBM Gene Expression")
drawHeatmap(GeneExp,group,silhouette=silhouette,scale="max_min",
            color="-RdYlBu",Title="GBM Gene Expression")
}
\author{
Xu,Taosheng \email{taosheng.x@gmail.com},Thuc Le \email{Thuc.Le@unisa.edu.au}
}
\references{
Gaujoux, Renaud, and Cathal Seoighe. "A flexible R package for nonnegative matrix factorization." BMC bioinformatics 11.1 (2010): 1.
}

