% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringMethod.R
\name{ExecuteiCluster}
\alias{ExecuteiCluster}
\title{Execute iCluster (Integrative clustering of multiple genomic data)}
\usage{
ExecuteiCluster(datasets, k, lambda = NULL, scale = TRUE, scalar = FALSE,
  max.iter = 10)
}
\arguments{
\item{datasets}{A list containing data matrices. For each data matrix, the rows represent genomic features, and the columns represent samples.
In order to unify the input parameter with other clustering methods, the data matrices are transposed comparing to the definition in the original "iCluster" package.}

\item{k}{Number of subtypes for the samples}

\item{lambda}{Penalty term for the coefficient matrix of the iCluster model}

\item{scale}{Logical value. If true, the genomic features in the matrix is centered.}

\item{scalar}{Logical value. If true, a degenerate version assuming scalar covariance matrix is used.}

\item{max.iter}{maximum iteration for the EM algorithm}
}
\value{
A list with the following elements.
\itemize{
 \item group  
 
 A vector represent the group of cancer subtypes. The order is corresponding to the samples in the data matrix.
 \item originalResult
 
 The clustering result of the original function "iCluster2()"
 }
}
\description{
Shen (2009) proposed a latent variable regression with a lasso constraint for joint modeling of multiple omics 
data types to identify common latent variables that can be used to cluster patient samples into biologically and clinically relevant disease subtypes.\cr
This function is based on the R package "iCluster". 
The R package "iCluster" should be installed. 
We write a shell to unify the input and output format.
It is helpful for the standardized flow of cancer subtypes analysis and validation. 
The parameters is compatible to the original R package "iCluster" function "iCluster2()".\cr
Please note: The data matrices are transposed in our function comparing to the original R package "iCluster" on the behalf of the unified input format with other functions.
We try to build a standardized flow for cancer subtypes analysis and validation.
}
\details{
For iCluster algorithm, it cannot process high-dimensional data, otherwise it is very very time-consuming or reports a mistake.  
 Based on test, it could smoothly run for the matrix with around 1500 features. Normally it need feature selection step first to reduce feature number.
}
\examples{
data(GeneExp)
data(miRNAExp)
data1=FSbyVar(GeneExp, cut.type="topk",value=500)
data2=FSbyVar(miRNAExp, cut.type="topk",value=100)
GBM=list(GeneExp=data1,miRNAExp=data2)
result=ExecuteiCluster(datasets=GBM, k=3, lambda=list(0.44,0.33,0.28))
result$group
}
\references{
Ronglai Shen, Adam Olshen, Marc Ladanyi. (2009). Integrative clustering of multiple genomic data types using a joint latent variable model with application to breast and lung cancer subtype analysis. Bioinformatics 25, 2906-2912.\cr
Ronglai Shen, Qianxing Mo, Nikolaus Schultz, Venkatraman E. Seshan, Adam B. Olshen, Jason Huse, Marc Ladanyi, Chris Sander. (2012). Integrative Subtype Discovery in Glioblastoma Using iCluster. PLoS ONE 7, e35236
}
\seealso{
\code{\link{iCluster2}}
}

