% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringMethod.R
\name{ExecuteSNF.CC}
\alias{ExecuteSNF.CC}
\title{Execute the combined SNF (Similarity Network Fusion) and Consensus clustering}
\usage{
ExecuteSNF.CC(
  datasets,
  clusterNum,
  K = 20,
  alpha = 0.5,
  t = 20,
  maxK = 10,
  pItem = 0.8,
  reps = 500,
  title = "ConsensusClusterResult",
  plot = "png",
  finalLinkage = "average"
)
}
\arguments{
\item{datasets}{A list containing data matrices. For each data matrix, 
the rows represent genomic features, and the columns represent samples. Same as ExecuteSNF}

\item{clusterNum}{A integer representing the return cluster number. Same as ExecuteSNF}

\item{K}{Number of nearest neighbors. Same as ExecuteSNF}

\item{alpha}{Variance for local model. Same as ExecuteSNF}

\item{t}{Number of iterations for the diffusion process. Same as ExecuteSNF}

\item{maxK}{integer value. maximum cluster number for Consensus Clustering Algorithm to evaluate. Same as ExecuteCC.}

\item{pItem}{Same as ExecuteCC}

\item{reps}{integer value. number of subsamples(in other words, The iteration number of each cluster number). Same as ExecuteCC}

\item{title}{character value for output directory. This title can be an absolute or relative path. Same as ExecuteCC}

\item{plot}{Same as ExecuteCC}

\item{finalLinkage}{Same as ExecuteCC}
}
\value{
Same as the ExecuteCC(). A list with the following elements.
\itemize{
 \item \strong{group} : A vector represent the group of cancer subtypes. The order is corresponding to the the samples in the data matrix.\cr
  
  This is the most important result for all clustering methods, so we place it as the first component. The format of group 
  is consistent across different algorithms and therefore makes it convenient for downstream analyses. Moreover, the format
  of group is also compatible with the K-means result and the hclust (after using the cutree() function).
  
 \item \strong{distanceMatrix} : It is a sample similarity matrix. The more large value between samples in the matrix, the more similarity the samples are.
  
  We extracted this matrix from the algorithmic procedure because it is useful for similarity analysis among the samples based on the clustering results.
  
 \item \strong{originalResult} : The clustering result of the original function "ConsensusClusterPlus()"
 
 Different clustering algorithms have different output formats. Although we have the group component which has consistent format for all of the algorithms (making it easy for downstream analyses), we still keep the output from the original algorithms.
 }
}
\description{
This function is a combined process of SNF and Consensus Clustering for cancer subtypes identification.
First it applied SNF to get the fusion patients similarity matrix. Then use this 
fusion patients similarity matrix as the sample distance for Consensus Clustering.
}
\examples{

data(GeneExp)
data(miRNAExp)
GBM=list(GeneExp,miRNAExp)
result=ExecuteSNF.CC(GBM, clusterNum=3, K=20, alpha=0.5, t=20,
                    maxK = 5, pItem = 0.8,reps=500, 
                    title = "GBM", plot = "png", 
                    finalLinkage ="average")
result$group
}
\seealso{
\code{\link{ExecuteSNF}} \code{\link{ExecuteCC}}
}
