% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{FSbyCox}
\alias{FSbyCox}
\title{Biological feature (such as gene) selection based on Cox regression model.}
\usage{
FSbyCox(Data, time, status, cutoff = 0.05)
}
\arguments{
\item{Data}{A data matrix representing the genomic data measured in a set of samples.
For the matrix, the rows represent the genomic features, and the columns represent the samples.}

\item{time}{A numeric vector representing the survival time (days) of a set of samples. Note that the
order of the time should map the samples in the Data matrix.}

\item{status}{A numeric vector representing the survival status of a set of samples. 0=alive/censored, 1=dead. Note that the
order of the time should map the samples in the Data matrix.}

\item{cutoff}{A numeric value in (0,1) representing whether the significant feature Xi is selected
according to the Proportional Hazards Assumption p-value of the feature Xi. If p-value(Xi)<cutoff, the
features Xi will be selected for downstream analysis. Normally the significant level is set to 0.05.}
}
\value{
A data matrix, extracted a subset with significant features from the input data matrix.
The rows represent the significant features, and the columns represents the samples.
}
\description{
Cox model (Proportional hazard model) is a statistical approach for survival risk analysis. We applied the univariate Cox model
for feature selection. The proportional hazard assumption test is used to evaluate the significant level of each biological feature
related to the survival result for samples. 
Eventually, the most significant genes are selected for clustering analysis.
}
\examples{
data(GeneExp)
data(time)
data(status)
data1=FSbyCox(GeneExp,time,status,cutoff=0.05)
}
\author{
Xu,Taosheng \email{taosheng.x@gmail.com}, Thuc Le \email{Thuc.Le@unisa.edu.au}
}
\references{
Andersen, P. and Gill, R. (1982). Cox's regression model for counting processes, a large sample study. Annals of Statistics 10, 1100-1120.\cr
Therneau, T., Grambsch, P., Modeling Survival Data: Extending the Cox Model. Springer-Verlag, 2000.
}

