% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{FSbyMAD}
\alias{FSbyMAD}
\title{Biological feature (such as gene) selection based on the most variant Median Absolute Deviation (MAD).}
\usage{
FSbyMAD(Data, cut.type = "topk", value)
}
\arguments{
\item{Data}{A data matrix representing the genomic data measured in a set of samples.
For the matrix, the rows represent the genomic features, and the columns represents the samples.}

\item{cut.type}{A character value representing the selection type. The optional values are shown below:
\itemize{
\item "topk"
\item "cutoff"
}}

\item{value}{A numeric value.\cr
If the cut.type="topk", the top number of value features are selected.\cr
If the cut.type="cutoff", the features with (MAD>value) are selected.}
}
\value{
An extracted subset data matrix with the most variant MAD features from the input data matrix.
}
\description{
Biological feature (such as gene) selection based on the most variant Median Absolute Deviation (MAD).
}
\examples{
data(GeneExp)
data1=FSbyMAD(GeneExp, cut.type="topk",value=1000)

}
\author{
Xu,Taosheng \email{taosheng.x@gmail.com}, Thuc Le \email{Thuc.Le@unisa.edu.au}
}

