% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{FSbyPCA}
\alias{FSbyPCA}
\title{Biological feature (such as gene) dimension reduction and extraction based on Principal Component Analysis.}
\usage{
FSbyPCA(Data, PC_percent = 1, scale = TRUE)
}
\arguments{
\item{Data}{A data matrix representing the genomic data measured in a set of samples.
For the matrix, the rows represent the genomic features, and the columns represents the samples.}

\item{PC_percent}{A numeric values in [0,1]  representing the ratio  of principal component is seclected.}

\item{scale}{A bool variable, If true, the Data is normalized before PCA.}
}
\value{
A new matrix with full or part Principal Component in new projection space.
}
\description{
This function is based on the prcomp(), we write a shell for it and make it easy to use on genomic data.
}
\examples{
data(GeneExp)
data1=FSbyPCA(GeneExp, PC_percent=0.9,scale = TRUE)
 
}
\author{
Xu,Taosheng \email{taosheng.x@gmail.com},Thuc Le \email{Thuc.Le@unisa.edu.au}
}

