% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneExp}
\alias{GeneExp}
\title{Dataset: Gene expression}
\format{A data matrix}
\description{
A glioblastoma (GBM) gene expression dataset downloaded from TCGA. This is a small dataset with 1500 genes 
and  100 cancer samples extracted from gene expression data for examples.
}
\details{
\itemize{
 \item Rows are genes
 \item Columns are cancer samples
}
}
\examples{
data(GeneExp)
}
\keyword{datasets}

