% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNAExp}
\alias{miRNAExp}
\title{Dataset: miRNA expression}
\format{A data matrix}
\description{
A glioblastoma (GBM) miRNA expression dataset downloaded from TCGA. This is a small miRNA expression dataset with 470 miRNAs and 
100 cancer samples extracted from miRNA expression data for examples.
}
\details{
\itemize{
 \item Rows are miRNAs
 \item Columns are cancer samples
}
}
\examples{
data(miRNAExp)
}
\keyword{datasets}

