% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valiation.R
\name{saveFigure}
\alias{saveFigure}
\title{This function save the figure in the current plot.}
\usage{
saveFigure(foldername = NULL, filename = "saveFig", image_width = 10,
  image_height = 10, image_res = 300)
}
\arguments{
\item{foldername}{Character values. It specifies the folder name which will be created in the present working path.}

\item{filename}{Character values. It specifies the saved file name.}

\item{image_width}{the figure width}

\item{image_height}{the figure height}

\item{image_res}{the figure resolution}
}
\value{
A * .png file in the specified folder.
}
\description{
This function save the figure in the current plot.
}
\examples{
data(GeneExp)
data(miRNAExp)
data(time)
data(status)
GBM=list(GeneExp=GeneExp,miRNAExp=miRNAExp)
result=ExecuteSNF(GBM, clusterNum=3, K=20, alpha=0.5, t=20)
group=result$group
distanceMatrix=result$distanceMatrix
p_value=survAnalysis(mainTitle="GBM",time,status,group,
      distanceMatrix=distanceMatrix,similarity=TRUE)
saveFigure(foldername="GBM",filename="GBM",image_width=10,image_height=10,image_res=300)
}
\author{
Xu,Taosheng \email{taosheng.x@gmail.com},Thuc Le \email{Thuc.Le@unisa.edu.au}
}

