% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valiation.R
\name{sigclustTest}
\alias{sigclustTest}
\title{A statistical method for testing the significance of clustering results.}
\usage{
sigclustTest(Data, group, nsim = 1000, nrep = 1, icovest = 1)
}
\arguments{
\item{Data}{A data matrix representing the genomic data measured in a set of samples.
For the matrix, the rows represent the genomic features, and the columns represents the samples.}

\item{group}{The subtypes label of each sample}

\item{nsim}{This is a parameter inherited from sigclust() in "sigclust" Package. 
Number of simulated Gaussian samples to estimate the distribution of the clustering index for the main p-value computation.}

\item{nrep}{This is a parameter inherited from sigclust() in "sigclust" Package. 
Number of steps to use in 2-means clustering computations (default=1, chosen to optimize speed).}

\item{icovest}{This is a parameter inherited from sigclust() in "sigclust" Package.
Covariance estimation type: 1. Use a soft threshold method as constrained MLE (default); 
2. Use sample covariance estimate (recommended when diagnostics fail); 3. Use original background noise threshold estimate (from Liu, et al, (2008)) ("hard thresholding").}
}
\value{
A matrix indicates the p-value between pairwise subtypes.
}
\description{
SigClust (Statistical significance of clustering) is a statistical method for testing the significance of clustering results. SigClust can be applied to
assess the statistical significance of splitting a data set into two clusters. 
SigClust studies whether clusters are really there, using the 2-means (k = 2) clustering index as a statistic. It assesses the significance of clustering by 
simulation from a single null Gaussian distribution. Null Gaussian parameters are estimated from the data.
Here we apply the SigClust to assess the statistical significance of pairwise subtypes. "sigclust" package should be installed.
}
\examples{
data(GeneExp)
data(miRNAExp)
data(time)
data(status)
GBM=list(GeneExp=GeneExp,miRNAExp=miRNAExp)
result=ExecuteSNF(GBM, clusterNum=3, K=20, alpha=0.5, t=20)
group=result$group
sigclust1=sigclustTest(miRNAExp,group, nsim=500, nrep=1, icovest=3)
sigclust2=sigclustTest(miRNAExp,group, nsim=1000, nrep=1, icovest=1)
}
\author{
Xu,Taosheng \email{taosheng.x@gmail.com},Thuc Le \email{Thuc.Le@unisa.edu.au}
}
\references{
Liu, Yufeng, Hayes, David Neil, Nobel, Andrew and Marron, J. S, 2008, Statistical Significance of Clustering for High-Dimension, Low-Sample Size Data, Journal of the American Statistical Association 103(483) 1281-1293.\cr
Huang, Hanwen, Yufeng Liu, Ming Yuan, and J. S. Marron. "Statistical Significance of Clustering Using Soft Thresholding." Journal of Computational and Graphical Statistics, no. just-accepted (2014): 00-00.
}
\seealso{
\code{\link{sigclust}}
}

