% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{spectralAlg}
\alias{spectralAlg}
\title{This is an internal function but need to be exported for the function ExecuteSNF.CC() call.}
\usage{
spectralAlg(affinity, K, type = 3)
}
\arguments{
\item{affinity}{Similarity matrix}

\item{K}{Number of clusters}

\item{type}{The variants of spectral clustering to use.}
}
\value{
A vector consisting of cluster labels of each sample.
}
\description{
This is  Spectral Clustering Algorithm extracted from SNFtools package spectralClustering() with
a tiny modification.
}
\examples{
####see the spectralClustering() in SNFtool package for the detail example.
data(miRNAExp)
Dist1=SNFtool::dist2(t(miRNAExp),t(miRNAExp))
W1 = SNFtool::affinityMatrix(Dist1, 20, 0.5)
group=spectralAlg(W1,3, type = 3)

}

