% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status}
\alias{status}
\title{Dataset: Survival status}
\format{A numeric vector}
\description{
\itemize{
 \item A vector representing the survival status for GBM cancer patients matched with the "GeneExp" and "miRNAExp"
. 0=alive or censored, 1=dead
}
}
\examples{
data(status)
}
\keyword{datasets}

