% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time}
\alias{time}
\title{Dataset: Survival time}
\format{A numeric vector}
\description{
\itemize{
 \item A vector representing the right censored survival time (days) for GBM cancer patients matched with the "GeneExp" and "miRNAExp"
datasets.
}
}
\examples{
data(time)
}
\keyword{datasets}

