% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valiation.R
\name{DiffExp.limma}
\alias{DiffExp.limma}
\title{DiffExp.limma}
\usage{
DiffExp.limma(Tumor_Data, Normal_Data, group = NULL, topk = NULL,
  RNAseq = FALSE)
}
\arguments{
\item{Tumor_Data}{A matrix representing the genomic data of cancer samples such as gene expression data, miRNA expression data.\cr
For the matrix, the rows represent the genomic features, and the columns represent the cancer samples.}

\item{Normal_Data}{A matrix representing the genomic data of Normal samples.\cr
For the matrix, the rows represent the genomic features corresponding to the Tumor_Data, and the columns represent the normal samples.}

\item{group}{A vector representing the subtype of each tumor sample in the Tumor_Data. The length of group is equal to the column number of Tumor_Data.}

\item{topk}{The top number of different expression features that we want to extract in the return result.}

\item{RNAseq}{A bool type representing the input datatype is a RNASeq or not. Default is FALSE for microarray data.}
}
\value{
A list representing the differently expression for each subtype comparing to the Normal group.
}
\description{
Differently Expression Analysis for genomic data. We apply limma package to conduct the analysis.
}
\examples{
data(GeneExp)
data(miRNAExp)
GBM=list(GeneExp=GeneExp,miRNAExp=miRNAExp)
result=ExecuteSNF(GBM, clusterNum=3, K=20, alpha=0.5, t=20)
group=result$group
######Fabricate a normal group by extracting some samples from the cancer dataset 
######for demonstrating the examples.
Normal_Data=GeneExp[,sample(1:100,20)]
result=DiffExp.limma(Tumor_Data=GeneExp,Normal_Data=Normal_Data,group=group,topk=NULL,RNAseq=FALSE)
}
\author{
Xu,Taosheng \email{taosheng.x@gmail.com},Thuc Le \email{Thuc.Le@unisa.edu.au}
}
\references{
Smyth, Gordon K. "Limma: linear models for microarray data." Bioinformatics and computational biology solutions using R and Bioconductor. 
Springer New York, 2005. 397-420.
}

