% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringMethod.R
\name{ExecuteSNF.CC}
\alias{ExecuteSNF.CC}
\title{Execute the combined SNF (Similarity Network Fusion) and Consensus clustering}
\usage{
ExecuteSNF.CC(datasets, clusterNum, K = 20, alpha = 0.5, t = 20,
  maxK = 10, pItem = 0.8, reps = 500, title = "ConsensusClusterResult",
  plot = "png", finalLinkage = "average")
}
\arguments{
\item{datasets}{Same as ExecuteSNF}

\item{clusterNum}{Same as ExecuteSNF}

\item{K}{Same as ExecuteSNF}

\item{alpha}{Same as ExecuteSNF}

\item{t}{Same as ExecuteSNF}

\item{maxK}{Same as ExecuteCC}

\item{pItem}{Same as ExecuteCC}

\item{reps}{Same as ExecuteCC}

\item{title}{Same as ExecuteCC}

\item{plot}{Same as ExecuteCC}

\item{finalLinkage}{Same as ExecuteCC}
}
\value{
Same as the ExecuteCC(). A list with the following elements.
\itemize{
 \item group
 
  A vector represent the group of cancer subtypes. The order is corresponding to the samples in the data matrix.
 \item distanceMatrix
 
  It is a sample dissimilarity matrix. The more large value between samples in the matrix, the more dissimilarity the samples are.
 \item originalResult
 
 The clustering result of the original function "ConsensusClusterPlus()"
 }
}
\description{
This function is a combined process of SNF and Consensus Clustering for cancer subtypes identification.
First it applied SNF to get the fusion patients similarity matrix. Then use this 
fusion patients similarity matrix as the sample distance for Consensus Clustering.
}
\examples{

data(GeneExp)
data(miRNAExp)
GBM=list(GeneExp,miRNAExp)
result=ExecuteSNF.CC(GBM, clusterNum=3, K=20, alpha=0.5, t=20,
                    maxK = 5, pItem = 0.8,reps=500, 
                    title = "GBM", plot = "png", 
                    finalLinkage ="average")
result$group
}
\seealso{
\code{\link{ExecuteSNF}} \code{\link{ExecuteCC}}
}

